make modules for rest api hierarchy;
