+++ bindings/ergo-lib-c-core/src/rest/node_info.rs
@@ -1,0 +1,3 @@
+#[derive(derive_more::From, derive_more::Into)]
+pub struct NodeInfo(pub(crate) ergo_lib::ergo_rest::NodeInfo);
+pub type NodeInfoPtr = *mut NodeInfo;
