+++ bindings/ergo-lib-c/Cargo.toml
@@ -15,6 +15,5 @@ exclude = [
 crate-type = ["staticlib"]
 
 [dependencies]
-ergo-lib = { version = "^0.14.0", path = "../../ergo-lib" , features = ["rest"]}
 ergo-lib-c-core = {version = "^0.14.0", path = "../ergo-lib-c-core"}
 paste = "^1.0"
