+++ bindings/ergo-lib-c/src/rest/api.rs
@@ -1,3 +1,4 @@
 //! REST API for the services in Ergo ecosystem (node, explorer, etc.)
 
 pub mod node;
+pub mod node_async;
